refactor(spells): enforce one-export-per-file and add localization
- Add localization keys for spell select/deselect tooltip and aria-label
- Split getSpellsFromIndex into separate module from getSpells.ts
- Inline sortSpellsBySchoolThenName where used (SchoolSelection, SpellSelection)
- Refactor spellDescription.ts to single default export (enrichSpellText)
- Move combineSpellDescriptionParts and fetchSpellDescriptionByUuid to callers
- Replace relative imports with path aliases throughout
