+++ src/view/dialogs/components/characterCreator/SpellSelection.svelte.ts
@@ -1,6 +1,16 @@
 import type { SpellIndexEntry } from '#utils/getSpells.js';
-import { sortSpellsBySchoolThenName } from '#utils/getSpells.js';
 import type { SpellSelectionGroup } from '../../characterCreation/types.js';
+
+/**
+ * Sorts spells by school (alphabetically) then by name (alphabetically).
+ */
+function sortSpellsBySchoolThenName(spells: SpellIndexEntry[]): SpellIndexEntry[] {
+	return [...spells].sort((a, b) => {
+		const schoolCompare = a.school.localeCompare(b.school);
+		if (schoolCompare !== 0) return schoolCompare;
+		return a.name.localeCompare(b.name);
+	});
+}
 
 export interface SpellSchoolGroup {
 	school: string;
